Finished moving code to FusedKernel files, fixed a bug in a unit test, and reduced the amount of warnings.

+++ include/fused_kernel/ptr_nd.cuh
@@ -247,7 +247,7 @@ struct PtrImpl<_2D,T> {
         if (ptr_a.dims.pitch == 0) {
             size_t pitch;
             gpuErrchk(cudaMallocPitch(&ptr_a.data, &pitch, sizeof(T) * ptr_a.dims.width, ptr_a.dims.height));
-            ptr_a.dims.pitch = pitch;
+            ptr_a.dims.pitch = (int)pitch;
         } else {
             gpuErrchk(cudaMalloc(&ptr_a.data, PtrImpl<_2D,T>::sizeInBytes(ptr_a.dims)));
         }
@@ -424,7 +424,7 @@ public:
         return deviceID;
     }
 
-    __host__ inline constexpr uint sizeInBytes() const {
+    __host__ inline constexpr size_t sizeInBytes() const {
         return PtrImpl<D, T>::sizeInBytes(ptr_a.dims);
     }
 
@@ -434,7 +434,7 @@ public:
 
     __host__ inline constexpr void setTo(const T& val) {
         if (type == MemType::Host || type == MemType::HostPinned) {
-            for (int i = 0; i < getNumElements(); i++) {
+            for (int i = 0; i < (int)getNumElements(); i++) {
                 ptr_a.data[i] = val;
             }
         } else {
